Moved files into more appropriate locations
