+++ src/domain/session/room/timeline/MessageBodyBuilder.js
